updated demo files and supporting resources to current version

+++ src/main/java/dqm/jku/trustkg/quality/profilingmetrics/singlecolumn/datatypeinfo/Decimals.java
@@ -55,7 +55,7 @@ public class Decimals extends ProfileMetric {
     if (decimals == -1) this.setValue(null);
     else {
     	this.setValue(decimals);
-    	this.setNumericVal(((Number) decimals).longValue());
+    	this.setNumericVal(((Number) decimals).intValue());
     }
   }
